added sample code to backbone.sync socketio implementation
The main problem with this implementation is that backbone.sync uses a
success callback in the options param to trigger the standard parsing
behavior, removing a successfully deleted model from a collection, etc.
 This callback then needs to get called in the socket.on(“…”, foo)
callback.  But socket.on can only be executed once, otherwise the
callbacks start stacking because they get registered every time.  This
is challenging.  A much easier and less over-ridingly invasive approach
is to hook into the repository, and send realtime events out each time
something of importance changes.  Let backbone use the ajax structure
that works so well against the standard API, and simply send out
messages when new models are created.